+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.7 (24/07/2016)
+- Added 'Called Times' feature sets
+- More Bug fixes and fixing failing tests
+
 # 0.2.5/0.2.6 (16/05/2016)
 - More Bug fixes and fixing failing tests
 
@@ -11,3 +15,4 @@
 # 0.2.2 (02/05/2016)
 - Updated Blog Section
 - Initial Release
+